+++ CHANGELOG.md
@@ -19,7 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
--   Handle new `wrapperClassnames` property on `DataList.Container` component to simplify implementation of custom
+-   Handle new `layoutClassnames` property on `DataList.Container` component to simplify implementation of custom
     `DataList` layouts
 
 ## [1.0.33] - 2024-06-06